refactor(tui): split interaction scroll methods into scroll.rs (400-line cap)
#988's scroll additions pushed src/tui/screens/interaction/mod.rs to 426 lines,
over the CI 400-line cap (File Size Lint). Move scroll_up/scroll_down/page_up/
page_down/jump_to_latest into a new scroll.rs submodule (same pattern as
layout.rs/turn.rs). mod.rs is now 389 lines. No behavior change.

Refs #988
